% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecd.R
\name{oecd_dataflows_popular}
\alias{oecd_dataflows_popular}
\title{List popular OECD dataflows}
\usage{
oecd_dataflows_popular()
}
\value{
tibble: short_name, agency, dataflow, description
}
\description{
Returns a curated list of commonly used OECD dataflows.
}
